Proposed pattern for getting multiple byte codes, not for check-in.
The change is a proposed pattern for getting multiple byte codes by the runtime that will provide a small performance improvement and reduced the lines of code.   Plus a bonus validation check as VOID is not valid for these operations.     The goal is pattern approval before making a large change.     Validated under Windows/msvc and Linux/g++